requirements.txt
<<<<<<< HEAD
Flask==2.3.3
flask-socketio==5.3.5
python-dotenv==1.0.0
paramiko==3.3.1
anthropic==0.5.0
SQLAlchemy==2.0.20
cryptography==41.0.3
eventlet==0.33.3
python-engineio==4.5.1
python-socketio==5.8.0 
=======
flask==2.0.1
flask-socketio==5.1.1
paramiko==2.8.1
anthropic==0.5.0
python-dotenv==0.19.1
pysqlite3==0.4.6
requests==2.26.0
gevent==21.8.0
gevent-websocket==0.10.1
>>>>>>> origin/main
